backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports-datetime-fromisoformat==2.0.3
click==8.2.1
ecdsa==0.19.1
environs==14.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
flake8==7.3.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
marshmallow==4.0.0
mccabe==0.7.0
passlib==1.7.4
pyasn1==0.6.1
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
